Redefine metadata model
